requirements.txt
confluent-kafka==2.3.0
pydantic==2.5.3
pydantic-settings==2.1.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
structlog==24.1.0
tenacity==8.2.3
schedule==1.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-httpx==0.28.0
websockets==12.0
python-multipart==0.0.6
